style.css fixed
I removed the navbar you did cause you put it as div, you should do the navbar as <nav></nav> and put the tag in a <header></header> tag cause it's better and more organized and I removed the css of the logo.
also use the images file and don't create another one like what you did Zeyad, also please change the background picture it doesn't fit and it's a game picture not a real thing you should use another one that is related to the contact us page.